updated ansible core version to 2.16 and code cleanup (#22)
* updated ansible core version to 2.16

* updated chanegelog

* updated pytho version

* fixed sanity tests

* fixed sanity tests

* replaced usuage of subprocess:

* clean up code

* clean up dependncies

* clean up code

* Update changelogs/fragments/ansible_core.yaml


---------

+++ plugins/module_utils/spec.py
@@ -13,16 +13,15 @@ import importlib
 import json
 import os
 import shutil
-import subprocess
 import sys
 import uuid
 
 from copy import deepcopy
 
 from ansible.module_utils.basic import missing_required_lib
-from ansible.module_utils.six import StringIO
+from io import StringIO
 
-from ansible_collections.ansible.yang.plugins.module_utils.common import find_file_in_path, to_list
+from ansible_collections.ansible.yang.plugins.module_utils.common import find_file_in_path, load_from_source, to_list
 
 
 # try:
@@ -61,14 +60,15 @@ class GenerateSpec(object):
         except Exception as exc:
             raise ValueError(missing_required_lib("pyang")) from exc
 
+        self._pyang_module = load_from_source(self._pyang_exec_path, "pyang")
         self._tmp_dir_path = tmp_dir_path
         self._handle_yang_file_path(yang_file_path)
         self._handle_search_path(search_path)
 
     def __del__(self):
-        if not self._keep_tmp_files:
-            shutil.rmtree(self._tmp_dir_path, ignore_errors=True)
-        super(GenerateSpec, self).__del__()
+        if hasattr(self, '_keep_tmp_files') and not self._keep_tmp_files:
+            if hasattr(self, '_tmp_dir_path'):
+                shutil.rmtree(self._tmp_dir_path, ignore_errors=True)
 
     def _handle_yang_file_path(self, yang_files):
         if not yang_files:
@@ -134,7 +134,7 @@ class GenerateSpec(object):
         )
         tree_tmp_file_path = os.path.realpath(os.path.expanduser(tree_tmp_file_path))
         # fill in the sys args before invoking pyang to retrieve tree structure
-        tree_cmd = [
+        sys.argv = [
             self._pyang_exec_path,
             "-f",
             "tree",
@@ -146,11 +146,7 @@ class GenerateSpec(object):
         ] + self._yang_file_path
 
         try:
-            subprocess.check_output(
-                " ".join(tree_cmd),
-                stderr=subprocess.STDOUT,
-                shell=True,
-            )
+            self._pyang_module.run()
         except SystemExit:
             pass
         except Exception as e:
@@ -159,7 +155,7 @@ class GenerateSpec(object):
                     os.path.realpath(os.path.expanduser(self._tmp_dir_path)),
                     ignore_errors=True,
                 )
-            raise ValueError("Error while generating skeleton xml file: %s" % e.output)
+            raise ValueError("Error while generating skeleton xml file: %s" % str(e))
         finally:
             err = sys.stdout.getvalue()
             if err and "error" in err.lower():
@@ -219,7 +215,7 @@ class GenerateSpec(object):
         )
         xml_tmp_file_path = os.path.realpath(os.path.expanduser(xml_tmp_file_path))
         # fill in the sys args before invoking pyang to retrieve xml skeleton
-        sample_xml_skeleton_cmd = [
+        sys.argv = [
             self._pyang_exec_path,
             "-f",
             "sample-xml-skeleton",
@@ -233,17 +229,13 @@ class GenerateSpec(object):
         ] + self._yang_file_path
 
         if defaults:
-            sample_xml_skeleton_cmd.append("--sample-xml-skeleton-defaults")
+            sys.argv.append("--sample-xml-skeleton-defaults")
 
         if annotations:
-            sample_xml_skeleton_cmd.append("--sample-xml-skeleton-annotations")
+            sys.argv.append("--sample-xml-skeleton-annotations")
 
         try:
-            subprocess.check_output(
-                " ".join(sample_xml_skeleton_cmd),
-                stderr=subprocess.STDOUT,
-                shell=True,
-            )
+            self._pyang_module.run()
         except SystemExit:
             pass
         except Exception as e:
@@ -252,7 +244,7 @@ class GenerateSpec(object):
                     os.path.realpath(os.path.expanduser(self._tmp_dir_path)),
                     ignore_errors=True,
                 )
-            raise ValueError("Error while generating skeleton xml file: %s" % e.output)
+            raise ValueError("Error while generating skeleton xml file: %s" % str(e))
         finally:
             err = sys.stdout.getvalue()
             if err and "error" in err.lower():
@@ -312,7 +304,7 @@ class GenerateSpec(object):
         )
         shutil.copy(plugin_file_src, self._tmp_dir_path)
         # fill in the sys args before invoking pyang to retrieve json skeleton
-        sample_json_skeleton_cmd = [
+        sys.argv = [
             self._pyang_exec_path,
             "--plugindir",
             self._tmp_dir_path,
@@ -328,14 +320,10 @@ class GenerateSpec(object):
         ] + self._yang_file_path
 
         if defaults:
-            sample_json_skeleton_cmd.append("--sample-json-skeleton-defaults")
+            sys.argv.append("--sample-json-skeleton-defaults")
 
         try:
-            subprocess.check_output(
-                " ".join(sample_json_skeleton_cmd),
-                stderr=subprocess.STDOUT,
-                shell=True,
-            )
+            self._pyang_module.run()
         except SystemExit:
             pass
         except Exception as e:
@@ -344,7 +332,7 @@ class GenerateSpec(object):
                     os.path.realpath(os.path.expanduser(self._tmp_dir_path)),
                     ignore_errors=True,
                 )
-            raise ValueError("Error while generating skeleton json file: %s" % e.output)
+            raise ValueError("Error while generating skeleton json file: %s" % str(e))
         finally:
             err = sys.stdout.getvalue()
             if err and "error" in err.lower():
